+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## `1.0.73-SNAPSHOT`
+## `1.0.86-SNAPSHOT`
 
 Release date `UNRELEASED`
+
+## `1.0.85-rc3`
+
+Release date `2023-02-27`
 
 - Implement `types/as-websocket-callback` using protocols for better 
   performance.